backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.110.0
gotrue==2.9.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.25.2
hyperframe==6.1.0
idna==3.11
packaging==25.0
postgrest==0.13.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.1
supafunc==0.3.3
typing_extensions==4.15.0
uvicorn==0.27.1
websockets==12.0
